+++ docs/changes.md
@@ -1,5 +1,8 @@
 # 最新动态
 
+2026/8/10
+  * object_workflow_cmd support clone(感谢兆坤提供补丁)
+
 2026/8/9
   * rename object_orchestrator to object_workflow_cmd (感谢兆坤提供补丁)